Extend all Encodable types with JSON/property list methods.
This isn't as type-safe as I'd like to be (even though it will
throw if the encoded type cannot be converted), but it allows
types that encode to bridged structs (e.g. [String:String]) to
be encoded to and from those formats.

This also allows there to be only one implementation of the
extensions, which is good.