Add options parameter to JSON decoding.
This allows clients to specify AllowFragments - the mutable
containers/leaves options would also technically be allowed,
but they're largely meaningless in this case.

However, I favor this over adding a boolean parameter to
match more closely with the encoding API, and to allow any
future options added to NSJSONSerialization to just work.